fix: Resolve ctrl-d not working problem

+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/15 15:39:14 by sanghupa          #+#    #+#             */
-/*   Updated: 2023/09/22 22:24:16 by sanghupa         ###   ########.fr       */
+/*   Updated: 2023/09/23 12:04:01 by sanghupa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -271,7 +271,7 @@ void	ft_unset(t_sent *node, t_elst *lst);
 
 /// readcmd
 /* src/readcmd/readcmd.c */
-size_t	readcmd(char *cmd, int debug_mode);
+int		readcmd(char *cmd, int debug_mode);
 
 /// parsecmd
 /* src/parsecmd/parsecmd.c */
